% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_and_graphics.R
\name{plot.species}
\alias{plot.species}
\title{Plot species level trait data}
\usage{
\method{plot}{species}(df, trait)
}
\arguments{
\item{df}{a data frame of TRY data that has been read into the workspace with
the \code{read.try()} function}

\item{trait}{the specific trait to plot by species}
}
\value{
a plot of species data
}
\description{
\code{plot.species} plots species data
}
\details{
This function plots species data
}
\keyword{graphics}
\keyword{plot}
