% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_and_graphics.R
\name{plot.traits}
\alias{plot.traits}
\title{Plot trait data by species}
\usage{
\method{plot}{traits}(df, species.list, trait.list, group)
}
\arguments{
\item{df}{a data frame of TRY data that has been read into the workspace with
the \code{read.try()} function}

\item{species.list}{the specific trait to plot by species}

\item{trait.list}{vector of traits to look at}

\item{group}{i don't remember what it does}
}
\value{
a plot of species data
}
\description{
\code{plot.species} plots species data
}
\details{
This function plots species data
}
\keyword{graphics}
\keyword{plot}
