% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion_and_cleaning.R
\name{read.try}
\alias{read.try}
\title{This file containts ingestion and formatting scripts
This function imports raw TRY data}
\usage{
read.try(data.path)
}
\arguments{
\item{data.path}{the path to the raw TRY data to be inputed}
}
\value{
creates a data frame containing the columns "DatasetID", "SpeciesName",
"ObservationID", "ObsDataID", "TraitID", "DataID", "ValueKindName", "Replicates",
"StdValue", "StdUnit", "ErrorRisk"
}
\description{
\code{read.try} brings in the data
}
\examples{
\dontrun{
read.try(system.file("extdata", "try_sample_data.txt", package = "TRYqa"))
}

}
\keyword{csv}
\keyword{data}
\keyword{input}
\keyword{read}
