% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum.traits}
\alias{sum.traits}
\title{Create trait summaries}
\usage{
\method{sum}{traits}(df, cwm, weights, species, traits)
}
\arguments{
\item{df}{a data frame of TRY data that has been read into the workspace with
the \code{read.try()} function}

\item{cwm}{a TRUE FALSE}

\item{weights}{a vector of weights for community weighted means. must be same length as species list}

\item{species}{a vectpr of species you'd like to look at}

\item{traits}{a vector of traits to look at}
}
\value{
creates a data frame contain the mean, median and standard deviation
by species for each trait included
}
\description{
\code{sum.traits} creates summary tables of traits.
}
\details{
This trait summarizes things
}
\keyword{summary}
\keyword{trait}
