% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_and_graphics.R
\name{theme_try}
\alias{theme_try}
\title{Read in TRY data}
\usage{
theme_try()
}
\description{
\code{theme_try} provides a plotting theme for TRY data
}
\details{
A plotting theme to be added to a \code{ggplot2} graph and base theme used in the
\code{TRYqa} package plotting functions
}
\keyword{plot}
\keyword{theme}
