% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRYqa-package.R
\docType{package}
\name{TRYqa-package}
\alias{TRYqa}
\alias{TRYqa-package}
\title{TRYqa: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jeff Atkins \email{jeffrey.atkins@usda.gov} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Elizabeth Agee (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Marcos Longo (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Anthony Walker (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
