% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion_and_cleaning.R, R/utilities.R
\name{read.try}
\alias{read.try}
\title{This file containts ingestion and formatting scripts
This function imports raw TRY data}
\usage{
read.try(data.path)

read.try(data.path)
}
\arguments{
\item{data.path}{the path to the raw TRY data to be inputed}
}
\value{
creates a data frame containing the columns "DatasetID", "SpeciesName",
"ObservationID", "ObsDataID", "TraitID", "DataID", "ValueKindName", "Replicates",
"StdValue", "StdUnit", "ErrorRisk"

creates a data frame containing the columns "DatasetID", "SpeciesName",
"ObservationID", "ObsDataID", "TraitID", "DataID", "ValueKindName", "Replicates",
"StdValue", "StdUnit", "ErrorRisk"
}
\description{
\code{read.try} brings in the data

\code{read.try} imports raw TRY data into the R environment.
}
\details{
This function imports raw TRY data
}
\examples{
read.try(system.file("extdata", "try_sample_data.txt", package = "TRYqa"))


# Run process complete PCL transect without storing to disk
data.path <- system.file("extdata", "try_sample_data.txt", package = "TRYqa")

read.try(data.path)

}
\keyword{csv}
\keyword{data}
\keyword{input}
\keyword{read}
